Fix index error in grid_test
In grid_test.py, function test_linefull1() would fail as it referenced row 24 (ie. a 25th row that does not exist) and get an index error for referencing an out of bounds area in the array. This fix changes those references to row 23 (ie. the 24th or final row that does exist). All 10 tests now succeed and none of them should fail.